requirements.txt
aggdraw==1.3.15
aiogram==2.21
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
Automat==20.2.0
autopep8==1.6.0
Babel==2.9.1
beautifulsoup4==4.11.1
boto3==1.21.37
botocore==1.24.37
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.12
click==7.1.2
click-plugins==1.1.1
cligj==0.7.2
config==0.5.1
constantly==15.1.0
cryptography==37.0.4
cssselect==1.1.0
cvxopt==1.3.0
cvxpy==1.2.1
cycler==0.11.0
distlib==0.3.5
docopt==0.6.2
ecos==2.0.10
emoji==2.0.0
filelock==3.7.1
Fiona==1.8.21
fonttools==4.33.3
frozenlist==1.3.0
future==0.18.2
GDAL==3.5.1
geopandas==0.10.2
HTMLParser==0.0.2
hyperlink==21.0.0
idna==3.3
imageio==2.19.3
incremental==21.3.0
itemadapter==0.6.0
itemloaders==1.0.4
jmespath==1.0.0
jsonschema==3.2.0
jsonseq==1.0.0
kiwisolver==1.4.2
lxml==4.9.1
mapbox-tilesets==1.7.3
matplotlib==3.5.2
maxrect @ git+https://github.com/planetlabs/maxrect.git@c91c240279cab57e77d348010c2683f944c653db
mercantile==1.1.6
multidict==6.0.2
munch==2.5.0
networkx==2.8.4
numpy==1.22.3
opencv-python==4.6.0.66
osqp==0.6.2.post5
packaging==21.3
pandas==1.4.2
parsel==1.6.0
Pillow==9.1.1
platformdirs==2.5.2
Protego==0.2.1
protobuf==3.19.4
psd-tools==1.9.21
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.9.2
pycodestyle==2.8.0
pycparser==2.21
PyDispatcher==2.0.5
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyproj==3.3.1
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
PyWavelets==1.3.0
qdldl==0.1.5.post2
queuelib==1.6.2
requests==2.27.1
requests-file==1.5.1
requests-toolbelt==0.9.1
s3transfer==0.5.2
scapy==2.4.5
scikit-image==0.19.3
scipy==1.8.1
Scrapy==2.6.2
scs==3.2.0
service-identity==21.1.0
Shapely==1.8.2
six==1.16.0
soupsieve==2.3.2.post1
telebot==0.0.4
tifffile==2022.5.4
tldextract==3.3.1
toml==0.10.2
Twisted==22.4.0
typing_extensions==4.3.0
urllib3==1.26.9
virtualenv==20.16.2
vk-api==11.9.7
w3lib==1.22.0
yarl==1.8.1
zope.interface==5.4.0
